TestAgent: Fail early if database is inaccessible (#366)
Make the start and restart commands of zonemaster_backend_testagent
fail before daemonization if the configured database is
inaccessible. This enables us to make service scripts that properly
report errors when the configured database engine isn't installed and
when there's a problem with the database connection configuration.

If the database connection fails early, zonemaster_backend_testagent
exits with a non-zero status and an error message to stderr.