Keep the "queue" parameter to compute fingerprint
The "queue" parameter has been kept because it is linked to a TestAgent
which can live in a specific environment. Using a pool of several
TestAgents in different testing environments each one with its own
queue, and adding the same test to each queue should result in the
creation of several tests (one per TestAgent in the pool). Then each
agent can run its own tests and produce its own results. The outcomes
might differ from on TestAgent to another.

Since the fingerprint is calculated with parameters that may affect the
outcome of a test, then the "queue" should be used in this computation.

The "priority" parameter however does not impact the outcome of a test.