New index to speed up query to retrieve next test
Looking for the next test to run involves searching among the tests
which haven't started (progress = 0) and sorting them by descending
priority. On huge "test_results" table (millions of entries) this query
can take several seconds. Using an index here can drastically speed up
the query to milleseconds (measurements made with PostgreSQL).

This new index is only added for PostgreSQL.

For SQLite, due to performance issue, it is not recommended to use it
for installation with heavy load (such as batch function).

When it comes to MariaDB and MySQL it seems that descending indexes were
not supported (and silently ignored [1]). MySQL 8 now supports
descending indexes [2]. However this is not yet the case for MariaDB
(there is a workaround [3]).
For this reason the index is not created for MariaDB/MySQL. If
performance is an issue, it is possible to create the index (when using
MySQL >= 8) or to use a workaround (involving a new PERSISTENT column
with reverse ordered values of priority, so that an ascending index can
be created on it).

[1]: https://jira.mariadb.org/browse/MDEV-13756
[2]: https://dev.mysql.com/doc/refman/8.0/en/mysql-nutshell.html
[3]: https://falseisnotnull.wordpress.com/2014/04/25/descending-indexes-in-mariadb/