Avoid SQL concatenation
Database driver allows "undef" bind parameters. However DBD::pg requires
a type to be passed. Therefore the `bind_param` method is used to pass
a SQL_INTEGER. In SQLite BOOLEAN is just an alias for an INTEGER with
the value 0 or 1.
To ease maintenance, the code is replicated across engines.