Include transitive test-time dependency on Test-Mojo
We depend on Mojoliciouse and JSON-Validator.
On Rocky there is a package for Mojolicious so we install that, but
there isn't one for JSON-Validator so we take that from CPAN (and run
its tests at installation-time).
JSON-Validator has a test-time dependency on Test::Mojo which is part of
the Mojolicious distribution.
However on Rocky Test::Mojo has been broken out from Mojolicious into
its own package.
So we need to install that too.