Properly trim trailing dots
Caution when the domain is only made of dots. For instance "...." should
result in "." and not an empty domain. To avoid this we first update any
domains equal to 2, 3 or 4 dots to a unique dot ".". Then we can safely
trim any trailing dots.
This is still not perfect though, because any domains only composed of
dots (at least 5 of them) will be normalized to the empty domain.