Compile JSON schemas before examining them
Related to #1105.

A subtle bug could cause RPCAPI methods, whose parameters were validated
against JSON schemas generated with JSON::Validator::Joi, to skip
validation of parameters against that schema entirely.

The offending code took a $method_schema, which could either be a raw
Perl hash containing a JSON schema or an opaque object made with
JSON::Validator::Joi, and tried to check whether the schema specified
any keys as required. It did so by checking whether the 'required' key
exists in $method_schema.

This works fine if $method_schema is a Perl hash. But when it’s a
JSON::Validator::Joi object, we are touching an implementation detail of
that class, which could sooner or later break!

And indeed, it broke: between JSON::Validator version 5.02 and 5.03, an
implementation detail was changed, causing issue #1105 on systems where
JSON::Validator version 5.03 or later is installed. The change seems to
have been introduced by commit 788fd42 [1] on JSON::Validator’s Github
repo [2].

[1]: https://github.com/jhthorsen/json-validator/commit/788fd4255899339d974d28141a5243f4409067d8
[2]: https://github.com/jhthorsen/json-validator