RPCAPI: check type before dereferencing hashref
Calling the RPCAPI endpoint with a number or a string instead of null or
an object caused a “500 Internal Server Error” to be returned, instead
of a propre RPCAPI error. This was due to the offending code failing to
check whether we were passed a hashref before attempting to dereference
it and checking the existence of an “id” key in it.