RPCAPI: fix logic error in JSON schema enforcement
There was another situation where JSON schema enforcement could be
skipped: when a method only has optional parameters and the “params” key
in the JSONRPC object was set to something, the value of that key was
not validated.

The offending code skipped validation altogether if none of the method’s
parameters is required. What was probably meant is that if none of the
method’s parameters are required, then the absence of a “params” key in
the JSONRPC object is acceptable. But a method requiring no parameters
doesn’t mean that it has no parameters at all (e.g. when a method has
default values for all its parameters), so validation should still be
done on a “params” key if it’s there.

The rewritten code implements a different, but more correct logic: if
the “params” key exists, validate its value and return an error if it
fails validation; if the “params” key does not exist and the method
requires at least one parameter, then return an error; else return
success.