PostgreSQL migration: avoid costly queries
The migration code for PostgreSQL caused some problems on a production
setup, where migrating a two-million-tuple test_results table failed to
complete despite 8 GB of RAM and 75 GB of free disk space.

It was assumed that PostgreSQL could complete this migration faster than
other DBMSes, owing to its native handling of JSON. This turned out to
be a double-edged sword.

For the time being, let’s work around the problem by dividing the
migration task into chunks of 50 000 tuples. It’s probably slower and
less elegant than the PostgreSQL-native solution, but at least it
doesn’t cause mayhem on the production machine.