PostgreSQL migration: use COPY for step 1
Instead of using INSERT statements, use COPY statements which have a
much higher throughput. The COPY data is generated on the fly using
the columns fetched from the server-side cursor.

The speedup for step 1 is approximately two-fold and on a copy of a
production system, takes migration time to two hours down from four
hours.