PostgreSQL migration: merge steps 1 and 3
Step 3 in the migration script for PostgreSQL could still cause some
problems with spikes in disk usage. It is also very inefficient,
because a full table scan is made on a table that has just been
populated.

In order to avoid that, the SELECT statement in step 1 is modified to
both populate the new table and migrate the Delegation01 messages in one
pass. The time taken for step 1 increases slightly (by about 10%) but
overall migration time should decrease further.