PostgreSQL migration: use json_to_recordset()
Instead of using json_array_elements() and then picking individual keys
in the thus obtained JSON (which is not very efficient), use
json_to_recordset() to turn each item of the results array into a table
with the appropriate type conversions already done for us. This should
net us a slight speed gain, as well as more legible SQL code.