PostgreSQL migration: run ANALYZE on log_level
In order to try and speed up step 1, I had a look at the output of
EXPLAIN applied to the query made at that step. The planner assumed that
the log_level table had 1,000 rows, which led it to optimize a join
poorly.

PostgreSQL has specific optimizations in place for joins on small
tables. But in order to benefit from that, PostgreSQL needs the
knowledge that log_level has only a handful of rows. Running ANALYZE
will make sure of that.