PostgreSQL migration: clean up cursors
This commit fixes one last issue with the PostgreSQL migration which
caused the final COMMIT to be expensive both in time and disk space.

It turns out we never needed to declare our cursor with “WITH HOLD”: the
entire migration is performed in one big transaction, which is committed
long after step 1 is done.

Declaring the cursor with “WITH HOLD” meant that committing the
migration to the database caused PostgreSQL to compute a copy of the
test_results table with the old data still there. Since the test_results
table was extensively modified, this copy takes up a significant amount
of disk space. Which is very wasteful, because the COMMIT is the last
thing being done in the migration, so that old version of test_results
required by the “WITH HOLD” cursor was thrown away immediately after
computation.

Without “WITH HOLD”, the cursor is closed when the final COMMIT is done.
But for good measure, we can close the cursor manually as soon as step 1
is done. Testing shows that the spike in disk usage no longer happens. I
also presume (although I wasn’t there to check) that the time taken by
that COMMIT operation is also significantly shorter.