Fix Backend-specific message tag translations
The recent case alteration of module names has caused messages whose
module attribute is BACKEND_TEST_AGENT to be stored as
Backend_Test_Agent instead. The database migration script for version
11.0 took care of that, but the rest of the code had not been adjusted
accordingly.

This meant that the Backend-specific TEST_DIED or UNABLE_TO_FINISH_TEST
tags were no longer translated appropriately.

In the meantime, new tests run on the backend will still store
BACKEND_TEST_AGENT instead of Backend_Test_Agent. To fix this
inconsistency, a change needs to be done to Zonemaster::Backend::DB and
a migration script needs to be written to fix the entries generated with
the backend between version 11.0 and the next release.