Rename Backend_Test_Agent to Backend in DB
For messages generated by the backend, store them as the “Backend”
module instead of “Backend_Test_Agent”. The use of mixed case with
underscores is unsatisfactory and the test agent is the only place in
the backend that generates these messages.

In essence, “Backend” could be a synonym for “System”, but it’s better
to keep the Engine’s system messages separate from the Backend’s.