Make sure we can store large test results
When storing large test results we construct an SQL statement with 7
variables per entry in the test result. Before DBD::SQLite 1.66 there
was a hard limit on 999 variables (SQLITE_LIMIT_VARIABLE_NUMBER). This
was only enough for small test results with up to 142 enries. In
DBD::SQLite 1.66 the limit was raised an we can store large test results
with up to 4680 entries.

N.B. This change doesn't solve the root cause. It only pushes the
problem in front of us by a good margin. IIRC the largest test result
we've observed was at about 2000 entries.